requirements.txt
streamlit
pandas
xgboost
shap
streamlit-shap
huggingface_hub
python-dotenv